Change ushort to uint. Change back matrix interface.  (#40)
* Abstract adjMatrix into Matrix. Implement SparseMatrix.

* Change matrix interface.

* Update SparseMatrix.hpp

* Fix indentation.

* Add SparseMatrix multiply function copyright.

* Update matrix interface when WEIGHTED is turned on

* remove redundant

* Change sparse matrix implementation

* Fix sparse matrix index usage.

* Change all ushort to uint.

* Rename VALUE_UNIT to WEIGHTED_VALUE.

* Add template support for class Matrix.

* Optimize initConnectedComponents function.

* Abstract CORES' data structure into Matrix.

* Disable printing measure 'ses' when SPARSE is enabled.

* Fix SparseMatrix multiply index.

* Change matrix's get set method back to brackets.

* Rename WEIGHTED_VALUE to MATRIX_UNIT.

+++ src/measures/localMeasures/EdgeCount.cpp
@@ -22,17 +22,17 @@ void EdgeCount::initSimMatrix() {
     uint n1 = G1->getNumNodes();
     uint n2 = G2->getNumNodes();
     uint k = distWeights.size();
-    vector<vector<ushort> > densities1 (n1, vector<ushort> (k+1));
-    vector<vector<ushort> > densities2 (n2, vector<ushort> (k+1));
-    for (ushort i = 0; i < n1; i++) {
+    vector<vector<uint> > densities1 (n1, vector<uint> (k+1));
+    vector<vector<uint> > densities2 (n2, vector<uint> (k+1));
+    for (uint i = 0; i < n1; i++) {
         densities1[i] = G1->numEdgesAround(i, k);
-        for (ushort j = 1; j < k; j++) {
+        for (uint j = 1; j < k; j++) {
             densities1[i][j] += densities1[i][j-1];
         } 
     }
-    for (ushort i = 0; i < n2; i++) {
+    for (uint i = 0; i < n2; i++) {
         densities2[i] = G2->numEdgesAround(i, k);
-        for (ushort j = 1; j < k; j++) {
+        for (uint j = 1; j < k; j++) {
             densities2[i][j] += densities2[i][j-1];
         }
     }
